# lite-backend/llm-config-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Config Gateway
  version: 0.1.0
servers:
  - url: http://localhost:9050
paths:
...
  /mcp/registry:
    get:
      tags: [MCP]
      summary: List registered MCP instances
      parameters:
        - in: query
          name: tenant
          schema: { type: string, default: default }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MCPInstance'
  /mcp/registry/register:
    post:
      tags: [MCP]
      summary: Register or update MCP instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPRegisterRequest'
      responses:
        '200': { description: OK }
...
  /mcp/registry/{tenant}/{name}:
    get:
      tags: [MCP]
      summary: Get MCP instance by tenant/name
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: string }
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPInstance'
        '404': { description: Not found }
    delete:
      tags: [MCP]
      summary: Delete MCP instance by tenant/name
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: string }
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
components:
  schemas:
    MCPInstance:
      type: object
      properties:
        id: { type: integer }
        tenant: { type: string }
        name: { type: string }
        server_name: { type: string }
        transport: { type: string, enum: [stdio, http, ws] }
        endpoint: { type: string }
        prefix: { type: string }
        status: { type: string, enum: [running, stopped, error] }
        tools_json: { type: string }
        metadata_json: { type: string }
        last_heartbeat: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MCPTool:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    MCPRegisterRequest:
      type: object
      required: [name, transport, endpoint]
      properties:
        tenant: { type: string, default: default }
        name: { type: string }
        server_name: { type: string }
        transport: { type: string, enum: [stdio, http, ws] }
        endpoint: { type: string }
        prefix: { type: string }
        status: { type: string }
        tools:
          type: array
          items:
            $ref: '#/components/schemas/MCPTool'
        metadata:
          type: object
          additionalProperties: true
...